Replace MongoDB to PostgreSQL with Exposed, update dependencies

+++ src/main/kotlin/com/github/yuppieflu/notifier/db/UserRepository.kt
@@ -6,5 +6,5 @@ import java.util.UUID
 interface UserRepository {
     fun persist(user: User): User
     fun findById(userId: UUID): User?
-    fun findByDeliveryHourAndEnabledSubscription(hour: Int): List<User>
+    fun findByDeliveryHourAndEnabledSubscription(hour: Byte): List<User>
 }
